chore: adds bookmark deletion error handling

+++ main/views.py
@@ -2,6 +2,7 @@ from django.contrib.auth.decorators import login_required
 from django.http import HttpResponseRedirect
 from django.shortcuts import render
 from django.urls import reverse
+from django.core.exceptions import ObjectDoesNotExist
 
 from .forms import BookmarkCreateForm
 from .models import Bookmark
@@ -30,12 +31,13 @@ def bookmarks(request):
 
 @login_required
 def bookmark_delete(request, id):
-    # FIXME error handling
-    bookmark = Bookmark.objects.get(id=id)
+    try:
+        bookmark = Bookmark.objects.get(id=id)
+    except ObjectDoesNotExist:
+        bookmark = None
 
-    if request.method == "POST":
-        success = bookmark.delete()
-        print(success)
+    if request.method == "POST" and bookmark is not None:
+        bookmark.delete()
         return HttpResponseRedirect(reverse("bookmarks"))
 
     return render(request, "bookmark_delete.html", {"bookmark": bookmark})
